related records: - display identity column value or referenced_column_name value for foreign keys; - format table value.
separate formatting function and reuse the same into table view and related records view.
