+++ frontend/src/app/components/db-table-row-edit/db-table-row-edit.component.html
@@ -64,7 +64,7 @@
         </h3>
 
         <mat-accordion multi="true">
-            <mat-expansion-panel *ngFor="let referencedTable of referencedTables; let i = index">
+            <mat-expansion-panel *ngFor="let referencedTable of referencedTables; let i = index" class="related-records__panel">
                 <mat-expansion-panel-header>
                     <mat-panel-title class="related-records__table-name"> {{referencedTable.displayTableName}} </mat-panel-title>
                     <mat-panel-description class="related-records__actions">
@@ -92,7 +92,7 @@
                         <span matListItemLine>
                             <span *ngFor="let field_name of referencedRecords[referencedTable.table_name].fieldsOrder">
                                 <strong>{{field_name}}:</strong>
-                                {{row[field_name] | json}}
+                                {{ row[field_name]}}
                             </span>
                         </span>
                     </a>
